+++ app/web.py
@@ -6,6 +6,7 @@ import json
 from flask import Flask
 from .api.pgs import Pgs
 from .api.transaction import Transaction
+from datetime import datetime
 
 class Web():
     SECRET_KEY = 'my_uno-tuti_secret_things'
@@ -19,6 +20,7 @@ class Web():
         self.flsk = Flask(__name__)
         self.flsk.config.from_object(self)
         self.trxs = []
+        self.tokens = []
 
     def run(self, config, logger):
         self.config = config
@@ -52,23 +54,91 @@ class Web():
             self.logger.info("Web Authentication success")
         return error
 
-    def get_shoppingCart(self, pp, lpn, amount):
-        """This method creates a new shopping cart for a tenant"""
+    def get_shoppingCart(self, pp, lpn, amount, tokenize):
+        """This method creates a new shopping cart for a tenant and can generate token"""
         self.logger.info(f"Web shopping cart for {pp} {lpn} amount {amount}")
         trx = Transaction(pp, lpn, amount)
-        trx = self.pgs.get_shopping_cart(trx)
+        if tokenize:
+            trx = self.pgs.get_shopping_cart(trx, True)
+        else:
+            trx = self.pgs.get_shopping_cart(trx, False)
+        try:
+            if trx.rsp_status_code == 200:
+                data = json.loads(trx.rsp_text)
+                for key in data:
+                    if key == 'cartId':
+                        trx.shoppingCartUuid = data[key]
+                        self.logger.info('Provided shopping cart {cartId}'.format(cartId=data[key])) 
+                    elif key == 'pgsToken':
+                        trx.pgsTokenUuid = data[key]
+                        self.logger.info('Provided token {token}'.format(token=data[key])) 
+            else:
+                if trx.rsp_status_code == 500:
+                    trx.rsp_status = str(trx.rsp_status_code)
+                    trx.rsp_code = "Internal Server Error"
+                else:
+                    data = json.loads(trx.rsp_text)
+                    for key in data:
+                        if key == 'code':
+                            trx.rsp_code = data[key]
+                        elif key == 'status':
+                            trx.rsp_status = data[key]
+        finally:
+            self.trxs.append(trx)
+        return trx
+
+    def get_tokenCart(self, pp, lpn, amount):
+        """This method creates a new token cart for a tenant"""
+        self.logger.info(f"Web token cart for {pp} {lpn} amount {amount}")
+        trx = Transaction(pp, lpn, amount)
+        trx = self.pgs.get_token_cart(trx)
+        try:
+            if trx.rsp_status_code == 200:
+                data = json.loads(trx.rsp_text)
+                for key in data:
+                    if key == 'cartId':
+                        trx.shoppingCartUuid = data[key]
+                        self.logger.info('Provided token cart {cartId}'.format(cartId=data[key])) 
+            else:
+                if trx.rsp_status_code == 500:
+                    trx.rsp_status = str(trx.rsp_status_code)
+                    trx.rsp_code = "Internal Server Error"
+                else:
+                    data = json.loads(trx.rsp_text)
+                    for key in data:
+                        if key == 'code':
+                            trx.rsp_code = data[key]
+                        elif key == 'status':
+                            trx.rsp_status = data[key]
+        finally:
+            self.trxs.append(trx)
+        return trx
+
+    def pay_tokenCart(self, trx, transInitiator):
+        """This method pay token cart for a tenant with using token"""
+        self.logger.info(f"Web pay token cart {trx.shoppingCartUuid} amount {trx.amount} with token {trx.pgsTokenUuid}")
+        trx = self.pgs.pay_token_cart(trx, trx.shop, trx.correlationId, trx.shoppingCartUuid, trx.pgsTokenUuid, transInitiator)
+
         if trx.rsp_status_code == 200:
             data = json.loads(trx.rsp_text)
             for key in data:
-                if key == 'cartId':
-                    trx.shoppingCartUuid = data[key]
-                    self.logger.info('Provided shopping cart {cartId}'.format(cartId=data[key])) 
-                    self.trxs.append(trx)
+                if key == 'status':
+                    trx.status = data[key]
+                elif key == 'payId':
+                    trx.trxId = data[key]
+                elif key == 'mediaType':
+                    trx.mediaType = data[key]
+                elif key == 'maskedMediaId':
+                    trx.maskedMediaId = data[key]
+                elif key == 'description':
+                    trx.description = data[key]
+            trx.author_time = datetime.now().strftime("%d.%m.%Y %H:%M:%S")
+            self.logger.info(f"TokenCart: {trx.shoppingCartUuid} Transaction: {trx.trxId} {trx.mediaType} {trx.maskedMediaId} {trx.status} Info:{trx.description}")
+
         else:
             if trx.rsp_status_code == 500:
                 trx.rsp_status = str(trx.rsp_status_code)
                 trx.rsp_code = "Internal Server Error"
-                self.trxs.append(trx)
             else:
                 data = json.loads(trx.rsp_text)
                 for key in data:
@@ -98,7 +168,11 @@ class Web():
                         trx.trx_imageUrls.append(method['imageUrl'])
                     else:
                         trx.trx_imageUrls.append("../static/none.png")
+                    if 'isTokenizationPossible' in method:
+                        trx.trx_possible_tokenization.append(method['isTokenizationPossible'])
+                    else:
+                        trx.trx_possible_tokenization.append("False")
 
             for id in range(len(trx.trx_methods)):
-                self.logger.info(f'{trx.trx_methods[id]} {trx.trx_fees[id]} {trx.trx_urls[id]} {trx.trx_imageUrls[id]}')
+                self.logger.info(f'{trx.trx_methods[id]} {trx.trx_fees[id]} {trx.trx_urls[id]} {trx.trx_imageUrls[id]} {trx.trx_possible_tokenization[id]}')
         return trx