+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-
+- added Pay & Tokenize feature
+- added Pay with Token feature (last generated token is used for payment)
 ### Changed
 
 ## [0.9.0](https://github.com/udigeri/inout/compare/v0.8.0..v0.9.0) (2021-12-01)